Updated done succefully !!!

+++ back/middleware/multer-config.js
@@ -22,25 +22,29 @@ const uploadAndCompressImage = (req, res, next) => {
     }
     // Utiliser le module sharp pour compresser l'image
     const timestamp = Date.now();
-    const name = `images/${timestamp}-${req.file.originalname.split(".")[0]}.webp`;
-    sharp(req.file.path)
-      .resize(800) // Redimensionner l'image à une largeur de 800 pixels
-      .jpeg({ quality: 80 }) // Compresser l'image en JPEG avec une qualité de 80%
-      .toFile(name, (err, info) => {
-        if (err) {
-          return res.status(500).json({ error: err.message });
-        }
-
-        // Supprimer l'image non compressée
-        fs.unlink(req.file.path, (err) => {
+    const name = req.file
+      ? `images/${timestamp}-${req.file.originalname.split(".")[0]}.webp`
+      : next();
+    if (req.file) {
+      sharp(req.file.path)
+        .resize(800) // Redimensionner l'image à une largeur de 800 pixels
+        .jpeg({ quality: 80 }) // Compresser l'image en JPEG avec une qualité de 80%
+        .toFile(name, (err, info) => {
           if (err) {
-            console.error(err);
+            return res.status(500).json({ error: err.message });
           }
+
+          // Supprimer l'image non compressée
+          fs.unlink(req.file.path, (err) => {
+            if (err) {
+              console.error(err);
+            }
+          });
+          req.file.name = name;
+          // Continuer la chaîne de middleware
+          next();
         });
-        req.file.name = name;
-        // Continuer la chaîne de middleware
-        next();
-      });
+    }
   });
 };
